+++ changelog_entry.yaml
@@ -1,5 +1,5 @@
 - bump: minor
   changes:
     added:
-    - Added employer side payroll liability variables.
-    - Added reform for taxing employer side payroll as employee contribution; in exchange, social security benefits are not taxed.
+    - Employer side Social Security and Medicare payroll tax.
+    - Reform for counting employer side payroll taxes in employees' IRS gross income.
