+++ policyengine_us/reforms/reforms.py
@@ -92,6 +92,9 @@ from .states.mt.ctc import (
 from .congress.golden import (
     create_fisc_act_reform,
 )
+from .tax_employer_payroll import(
+    create_tax_employer_payroll_reform
+)
 
 from policyengine_core.reforms import Reform
 import warnings
@@ -197,6 +200,7 @@ def create_structural_reforms_from_parameters(parameters, period):
     )
     mt_ctc = create_mt_ctc_reform(parameters, period)
     fisc_act = create_fisc_act_reform(parameters, period)
+    tax_employer_payroll = create_tax_employer_payroll_reform(parameters, period)
 
     reforms = [
         afa_reform,
@@ -237,6 +241,7 @@ def create_structural_reforms_from_parameters(parameters, period):
         nyc_school_tax_credit_with_phase_out,
         mt_ctc,
         fisc_act,
+        tax_employer_payroll,
     ]
     reforms = tuple(filter(lambda x: x is not None, reforms))
