+++ policyengine_us/reforms/tax_employer_payroll/__init__.py
@@ -1,0 +1,11 @@
+"""
+This makes the function create_tax_employer_payroll_reform available for 
+use without needing to specify the full module path.
+i.e., we can just call: 
+from policyengine_us.reforms.tax_employer_payroll import create_tax_employer_payroll_reform
+instead of:
+from policyengine_us.reforms.tax_employer_payroll.tax_employer_payroll import create_tax_employer_payroll_reform
+"""
+from .tax_employer_payroll import(
+    create_tax_employer_payroll_reform
+)
